Remove all the generated autotools files
All these files are generated by autotools, and thus there is no need
to store them statically in the repository (where they will only pile
up, growing the size of the repository). Also, the mismatch between the
version of autotools on the host vs the ones used to generated the
current files in the repository causes troubles.

Instead, let users (or make dist) generate them all, using a simple
run of autoreconf.  Add them to .gitignore, so there is no risk to
commit them again.

Fixes: #25